requirements.txt
fastapi == 0.104.1
openai == 0.27.8
SpeechRecognition == 3.10.0
asyncio == 3.4.3
python-dotenv == 1.0.0
pymongo == 4.6.0
passlib == 1.7.4
fastapi-jwt-auth == 0.5.0
pydantic == 1.10.7
fastapi-mail == 1.4.1
